refactor(frontend): the API-keys service moves into @agenta/settings
First settings extraction with a data half rather than presentation only. The
service was 46 lines over axios + the project id, both already in
@agenta/shared/api, so it moves with an import swap. OSS re-exports it, keeping
one implementation for its three consumers — one of which is in EE, so this is
also the first settings piece to cross editions.

The APIKeys page itself is untouched; splitting its fetch/permissions/dialogs
into a headless hook plus a view is the next unit.
