+++ web/packages/agenta-settings/src/api/apiKeys.ts
@@ -1,0 +1,36 @@
+import {axios, getAgentaApiUrl} from "@agenta/shared/api"
+import {getDefaultStore} from "jotai"
+
+import {projectIdAtom} from "@agenta/shared/state"
+
+const currentProjectId = () => getDefaultStore().get(projectIdAtom) ?? ""
+
+/**
+ * Workspace API keys.
+ *
+ * Prefix convention: fetchAll = GET all, create = POST, delete = DELETE.
+ * `ignoreAxiosError` suppresses the global error toast for callers that render their own.
+ */
+export const fetchAllListApiKeys = (workspaceId: string, ignoreAxiosError = false) =>
+    axios.get(
+        `${getAgentaApiUrl()}/keys/?workspace_id=${workspaceId}&project_id=${currentProjectId()}`,
+        {_ignoreError: ignoreAxiosError} as never,
+    )
+
+export const createApiKey = (
+    workspaceId: string,
+    ignoreAxiosError = false,
+    projectId?: string | null,
+) =>
+    axios.post(
+        `${getAgentaApiUrl()}/keys/?workspace_id=${workspaceId}&project_id=${
+            projectId ?? currentProjectId()
+        }`,
+        undefined,
+        {_ignoreError: ignoreAxiosError} as never,
+    )
+
+export const deleteApiKey = (prefix: string, ignoreAxiosError = false) =>
+    axios.delete(`${getAgentaApiUrl()}/keys/${prefix}?project_id=${currentProjectId()}`, {
+        _ignoreError: ignoreAxiosError,
+    } as never)
