+++ web/packages/agenta-settings/src/index.ts
@@ -5,3 +5,4 @@ export {
     useSettingsAccess,
     CLOSED_SETTINGS_ACCESS,
 } from "./access"
+export {fetchAllListApiKeys, createApiKey, deleteApiKey} from "./api/apiKeys"
